LayoutTests/platform/gtk/fast/viewport/viewport-82-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "400%" for key "width" was truncated to its numeric prefix.
viewport size 400x440 scale 0.800000 with limits [0.800000, 5.000000] and userScalable -1.000000
